Bug 1580266 - explicitly specify the default tree for help.html purposes. r=kats

+++ infrastructure/web-server-setup.sh
@@ -28,6 +28,8 @@ rm -rf $SERVER_ROOT/docroot
 mkdir -p $SERVER_ROOT/docroot
 DOCROOT=$(realpath $SERVER_ROOT/docroot)
 
+DEFAULT_TREE_NAME=$($MOZSEARCH_PATH/scripts/read-json.py $CONFIG_FILE default_tree)
+
 for TREE_NAME in $($MOZSEARCH_PATH/scripts/read-json.py $CONFIG_FILE trees)
 do
     mkdir -p $DOCROOT/file/$TREE_NAME
@@ -35,8 +37,13 @@ do
     ln -s $WORKING/$TREE_NAME/file $DOCROOT/file/$TREE_NAME/source
     ln -s $WORKING/$TREE_NAME/dir $DOCROOT/dir/$TREE_NAME/source
 
-    rm -f $DOCROOT/help.html
-    ln -s $WORKING/$TREE_NAME/help.html $DOCROOT
+    # Only update the help file if no default tree was specified OR
+    # The tree was specified and this is that tree.
+    if [ -z "$DEFAULT_TREE_NAME" -o "$DEFAULT_TREE_NAME" == "$TREE_NAME" ]
+    then
+        rm -f $DOCROOT/help.html
+        ln -s $WORKING/$TREE_NAME/help.html $DOCROOT
+    fi
 done
 
 python $MOZSEARCH_PATH/scripts/nginx-setup.py $CONFIG_FILE $DOCROOT "$USE_HSTS" > /tmp/nginx
